# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Inspection App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drives:
    get:
      operationId: listDrives
      tags: [drives]
      summary: List drives
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [venue, inspector] }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: venueId
          schema: { type: integer }
        - in: query
          name: holderUserId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: mine
          schema: { type: boolean }
      responses:
        "200":
          description: Drives
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DriveSummary" }
  /drives/{id}:
    get:
      operationId: getDrive
      tags: [drives]
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Drive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Drive" }
    patch:
      operationId: updateDrive
      tags: [drives]
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDriveInput" }
      responses:
        "200":
          description: Drive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Drive" }
  /drives/swap:
    post:
      operationId: swapDrive
      tags: [drives]
      summary: Swap a drive at a venue (extract and/or install)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SwapDriveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /drives/{id}/release:
    post:
      operationId: releaseDrive
      tags: [drives]
      summary: Maintenance releases a drive to an inspector
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReleaseDriveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /drives/{id}/accept:
    post:
      operationId: acceptDrive
      tags: [drives]
      summary: Accept a drive in transit
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /drives/{id}/return:
    post:
      operationId: returnDrive
      tags: [drives]
      summary: Inspector returns a drive to maintenance
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReturnDriveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /drives/whereabouts:
    get:
      operationId: driveWhereabouts
      tags: [drives]
      summary: Find drive holding footage for a venue + datetime
      parameters:
        - in: query
          name: venueId
          required: true
          schema: { type: integer }
        - in: query
          name: datetime
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Whereabouts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhereaboutsResult" }
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    DriveType:
      type: string
      enum: [venue, inspector]
    DriveSummary:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { $ref: "#/components/schemas/DriveType" }
        homeVenueId:
          type: ["integer", "null"]
        homeVenueName:
          type: ["string", "null"]
        homeVenueCode:
          type: ["string", "null"]
        status: { type: string }
        holderUserId:
          type: ["integer", "null"]
        holderName:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, type, homeVenueId, homeVenueName, homeVenueCode, status, holderUserId, holderName, createdAt, updatedAt]
    DriveFootageWindow:
      type: object
      properties:
        id: { type: integer }
        driveId: { type: integer }
        venueId: { type: integer }
        venueName: { type: string }
        venueCode: { type: string }
        installedAt: { type: string, format: date-time }
        extractedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, driveId, venueId, venueName, venueCode, installedAt, extractedAt]
    Drive:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { $ref: "#/components/schemas/DriveType" }
        homeVenueId:
          type: ["integer", "null"]
        homeVenueName:
          type: ["string", "null"]
        homeVenueCode:
          type: ["string", "null"]
        status: { type: string }
        holderUserId:
          type: ["integer", "null"]
        holderName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        footageWindows:
          type: array
          items: { $ref: "#/components/schemas/DriveFootageWindow" }
      required: [id, name, type, homeVenueId, homeVenueName, homeVenueCode, status, holderUserId, holderName, notes, createdAt, updatedAt, footageWindows]
    UpdateDriveInput:
      type: object
      properties:
        status: { type: string }
        notes:
          type: ["string", "null"]
    SwapDriveInput:
      type: object
      properties:
        venueId: { type: integer }
        installDriveId:
          type: ["integer", "null"]
        extractDriveId:
          type: ["integer", "null"]
      required: [venueId]
    ReleaseDriveInput:
      type: object
      properties:
        toUserId: { type: integer }
      required: [toUserId]
    ReturnDriveInput:
      type: object
      properties:
        toUserId: { type: integer }
      required: [toUserId]
    WhereaboutsMatch:
      type: object
      properties:
        windowId: { type: integer }
        driveId: { type: integer }
        driveName: { type: string }
        driveStatus: { type: string }
        holderUserId:
          type: ["integer", "null"]
        holderName:
          type: ["string", "null"]
        installedAt: { type: string, format: date-time }
        extractedAt:
          type: ["string", "null"]
          format: date-time
        likelyOverwritten: { type: boolean }
      required: [windowId, driveId, driveName, driveStatus, holderUserId, holderName, installedAt, extractedAt, likelyOverwritten]
    WhereaboutsInspection:
      type: object
      properties:
        id: { type: integer }
        dvrNumber: { type: string }
        footageDate: { type: string, format: date }
        driveId:
          type: ["integer", "null"]
      required: [id, dvrNumber, footageDate, driveId]
    WhereaboutsResult:
      type: object
      properties:
        matches:
          type: array
          items: { $ref: "#/components/schemas/WhereaboutsMatch" }
        inspections:
          type: array
          items: { $ref: "#/components/schemas/WhereaboutsInspection" }
      required: [matches, inspections]
...