# lib/api-zod/src/generated/types/drive.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Inspection App API
 * OpenAPI spec version: 0.1.0
 */
import type { DriveFootageWindow } from "./driveFootageWindow";
import type { DriveType } from "./driveType";

export interface Drive {
  id: number;
  name: string;
  type: DriveType;
  /** @nullable */
  homeVenueId: number | null;
  /** @nullable */
  homeVenueName: string | null;
  /** @nullable */
  homeVenueCode: string | null;
  status: string;
  /** @nullable */
  holderUserId: number | null;
  /** @nullable */
  holderName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  updatedAt: Date;
  footageWindows: DriveFootageWindow[];
}