# lib/api-zod/src/generated/types/driveSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Inspection App API
 * OpenAPI spec version: 0.1.0
 */
import type { DriveType } from "./driveType";

export interface DriveSummary {
  id: number;
  name: string;
  type: DriveType;
  /** @nullable */
  homeVenueId: number | null;
  /** @nullable */
  homeVenueName: string | null;
  /** @nullable */
  homeVenueCode: string | null;
  status: string;
  /** @nullable */
  holderUserId: number | null;
  /** @nullable */
  holderName: string | null;
  createdAt: Date;
  updatedAt: Date;
}